chore: Rename LDAP files to include .ee (no-changelog)
