+++ packages/cli/src/auth/methods/ldap.ts
@@ -1,7 +1,7 @@
 import { Container } from 'typedi';
 
 import { InternalHooks } from '@/InternalHooks';
-import { LdapService } from '@/Ldap/ldap.service';
+import { LdapService } from '@/Ldap/ldap.service.ee';
 import {
 	createLdapUserOnLocalDb,
 	getUserByEmail,
@@ -10,7 +10,7 @@ import {
 	mapLdapAttributesToUser,
 	createLdapAuthIdentity,
 	updateLdapUserOnLocalDb,
-} from '@/Ldap/helpers';
+} from '@/Ldap/helpers.ee';
 import type { User } from '@db/entities/User';
 import { EventRelay } from '@/eventbus/event-relay.service';
