+++ packages/cli/test/unit/Ldap/helpers.test.ts
@@ -1,6 +1,6 @@
 import { UserRepository } from '@/databases/repositories/user.repository';
 import { mockInstance } from '../../shared/mocking';
-import * as helpers from '@/Ldap/helpers';
+import * as helpers from '@/Ldap/helpers.ee';
 import { AuthIdentity } from '@/databases/entities/AuthIdentity';
 import { User } from '@/databases/entities/User';
 import { generateNanoId } from '@/databases/utils/generators';
